third_party/WebKit/LayoutTests/inspector/tracing/timeline-js/timeline-script-tag-1.html

<html>
<head>
<script src="../../../http/tests/inspector/inspector-test.js"></script>
<script src="../../../http/tests/inspector/timeline-test.js"></script>
<script>

function performActions()
{
    var iframe = document.createElement("iframe");
    iframe.src = "../resources/timeline-iframe-data.html";
    document.body.appendChild(iframe);
}

function test() 
{
    UI.panels.timeline._disableCaptureJSProfileSetting.set(true);
    PerformanceTestRunner.startTimeline(step1);
    function step1()
    {
        ConsoleTestRunner.addConsoleSniffer(step2);
        TestRunner.evaluateInPage("performActions()");
    }

    function step2()
    {
        PerformanceTestRunner.stopTimeline(step3);
    }

    function step3()
    {
        PerformanceTestRunner.timelineModel().mainThreadEvents().forEach(event => {
            if (event.name === TimelineModel.TimelineModel.RecordType.EvaluateScript) {
                PerformanceTestRunner.printTraceEventProperties(event);
            } else if (event.name === TimelineModel.TimelineModel.RecordType.TimeStamp) {
                TestRunner.addResult(`----> ${Timeline.TimelineUIUtils.eventTitle(event)}`);
            }
        });
        TestRunner.completeTest();
    }
}

</script>
</head>

<body onload="runTest()">
<p>
Tests the Timeline API instrumentation of an HTML script tag.
</p>

</body>
</html>
